requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.9.2
blas=1.0=mkl
box2d-py=2.3.5=pypi_0
brotli=1.0.9=h827c3e9_9
brotli-bin=1.0.9=h827c3e9_9
brotli-python=1.0.9=py311h5da7b33_9
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.12.31=haa95532_0
certifi=2024.12.14=py311haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
cloudpickle=3.1.0=pypi_0
contourpy=1.3.1=py311h214f63a_0
cuda-cccl=12.6.77=0
cuda-cccl_win-64=12.6.77=0
cuda-cudart=12.4.127=0
cuda-cudart-dev=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.1=0
cuda-libraries-dev=12.4.1=0
cuda-nvrtc=12.4.127=0
cuda-nvrtc-dev=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.6.77=0
cuda-opencl-dev=12.6.77=0
cuda-profiler-api=12.6.77=0
cuda-runtime=12.4.1=0
cuda-version=12.6=3
cycler=0.11.0=pyhd3eb1b0_0
farama-notifications=0.0.4=pypi_0
filelock=3.13.1=py311haa95532_0
fonttools=4.51.0=py311h2bbff1b_0
freetype=2.12.1=ha860e81_0
fsspec=2024.12.0=pypi_0
giflib=5.2.2=h7edc060_0
gmpy2=2.1.2=py311h7f96b67_0
gym-notices=0.0.8=pypi_0
gymnasium=1.0.0=pypi_0
icu=73.1=h6c2663c_0
idna=3.7=py311haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.4=py311haa95532_1
jpeg=9e=h827c3e9_3
keyboard=0.13.5=pypi_0
kiwisolver=1.4.4=py311hd77b12b_0
krb5=1.20.1=h5b6d351_0
lcms2=2.16=hb4a4139_0
lerc=4.0.0=h5da7b33_0
libbrotlicommon=1.0.9=h827c3e9_9
libbrotlidec=1.0.9=h827c3e9_9
libbrotlienc=1.0.9=h827c3e9_9
libclang=14.0.6=default_hb5a9fac_2
libclang13=14.0.6=default_h8e68704_2
libcublas=12.4.5.8=0
libcublas-dev=12.4.5.8=0
libcufft=11.2.1.3=0
libcufft-dev=11.2.1.3=0
libcurand=10.3.7.77=0
libcurand-dev=10.3.7.77=0
libcusolver=11.6.1.9=0
libcusolver-dev=11.6.1.9=0
libcusparse=12.3.1.170=0
libcusparse-dev=12.3.1.170=0
libdeflate=1.22=h5bf469e_0
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.2.5.30=0
libnpp-dev=12.2.5.30=0
libnvfatbin=12.6.77=0
libnvfatbin-dev=12.6.77=0
libnvjitlink=12.4.127=0
libnvjitlink-dev=12.4.127=0
libnvjpeg=12.3.1.117=0
libnvjpeg-dev=12.3.1.117=0
libpng=1.6.39=h8cc25b3_0
libpq=17.2=h70ee33d_0
libtiff=4.5.1=h44ae7cf_1
libuv=1.48.0=h827c3e9_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h3d04722_1
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.3=py311h827c3e9_1
matplotlib=3.10.0=py311haa95532_0
matplotlib-base=3.10.0=py311he19b0ae_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py311h827c3e9_2
mkl_fft=1.3.11=py311h827c3e9_0
mkl_random=1.2.8=py311hea22821_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py311haa95532_0
mss=10.0.0=pypi_0
networkx=3.2.1=py311haa95532_0
numpy=2.2.1=pypi_0
numpy-base=2.0.1=py311hd01c5d8_1
opencv-python=4.10.0.84=pypi_0
openjpeg=2.5.2=hae555c5_0
openssl=3.0.15=h827c3e9_0
packaging=24.2=py311haa95532_0
pillow=11.0.0=py311h096bfcc_1
pip=24.2=py311haa95532_0
ply=3.11=py311haa95532_0
pydirectinput=1.0.4=pypi_0
pygame=2.6.1=pypi_0
pyglet=2.1.0=pypi_0
pynput=1.7.7=pypi_0
pyparsing=3.2.0=py311haa95532_0
pyqt=5.15.10=py311hd77b12b_0
pyqt5-sip=12.13.0=py311h2bbff1b_0
pysocks=1.7.1=py311haa95532_0
pytesseract=0.3.13=pypi_0
python=3.11.11=h4607a30_0
python-dateutil=2.9.0post0=py311haa95532_2
pytorch=2.5.1=py3.11_cuda12.4_cudnn9_0
pytorch-cuda=12.4=h3fd98bf_7
pytorch-mutex=1.0=cuda
pywin32=308=pypi_0
pyyaml=6.0.2=py311h827c3e9_0
qt-main=5.15.2=h19c9488_11
requests=2.32.3=py311haa95532_1
setuptools=75.1.0=py311haa95532_0
sip=6.7.12=py311h5da7b33_1
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h2bbff1b_0
swig=4.3.0=pypi_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=h59b6b97_0
tk=8.6.14=h0416ee5_0
torchaudio=2.5.1=pypi_0
torcheval=0.0.7=pypi_0
torchvision=0.20.1=pypi_0
tornado=6.4.2=py311h827c3e9_0
typing_extensions=4.12.2=py311haa95532_0
tzdata=2024b=h04d1e81_0
unicodedata2=15.1.0=py311h827c3e9_1
urllib3=2.2.3=py311haa95532_0
vc=14.40=haa95532_2
vs2015_runtime=14.42.34433=h9531ae6_2
wheel=0.44.0=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
